ai-service/main.py
from fastapi import FastAPI
from pydantic import BaseModel
from groq import Groq
from dotenv import load_dotenv
import os
import json

load_dotenv()

app = FastAPI()

client = Groq(api_key=os.getenv("GROQ_API_KEY"))

class MessageRequest(BaseModel):
    username: str
    message: str
class ExamQuestionRequest(BaseModel):
    exam_type: str
    topic: str
    num_questions: int
    difficulty: str = "medium"   # optional: easy, medium, hard
class TrendRequest(BaseModel):
    course: str

@app.post("/ask")
async def ask_ai(request: MessageRequest):

    completion = client.chat.completions.create(
        model="groq/compound",
        messages = [
        {   
        "role": "system",
        "content": "You are a helpful assistant. Always respond in 2-3 simple bullet points or one short paragraph. Keep answers concise and easy to understand."
        },
        {
        "role": "user",
        "content": request.message
        }
        ]
        ,
        temperature=0.7,
        max_completion_tokens=1024,
        compound_custom={
            "tools": {
                "enabled_tools":["web_search","code_interpreter","visit_website"]
            }
        }
    )

    return {
        "reply": completion.choices[0].message.content
    }

@app.post("/chat")
async def chat_ai(request: MessageRequest):

    completion = client.chat.completions.create(
        model="groq/compound",
        messages = [
        {   
        "role": "system",
        "content": "You are a helpful assistant. Always respond in 6-7 simple bullet points . Keep answers concise and easy to understand."
        },
        {
        "role": "user",
        "content": request.message
        }
        ]
        ,
        temperature=0.7,
        max_completion_tokens=1024,
        compound_custom={
            "tools": {
                "enabled_tools":["web_search","code_interpreter","visit_website"]
            }
        }
    )

    return {
        "reply": completion.choices[0].message.content
    }
@app.post("/generate-questions")
async def generate_questions(request: ExamQuestionRequest):

    prompt = f"""
    Generate questions for the following exam:

    Exam Type: {request.exam_type}
    Topic: {request.topic}
    Number of Questions for practice: {request.num_questions}
    Difficulty: {request.difficulty}

    IMPORTANT: Respond ONLY with valid JSON. Do not include any markdown formatting, code blocks, or extra text.

    JSON structure:
    {{
        "practice_questions": [
            {{
                "question": "Question text here",
                "options": ["Option A", "Option B", "Option C", "Option D"],
                "correct_answer": "A",
                "solution": "Detailed solution explanation"
            }}
        ],
        "quiz_questions": [
            {{
                "question": "Question text here",
                "options": ["Option A", "Option B", "Option C", "Option D"],
                "correct_answer": "A",
                "solution": "Detailed solution explanation"
            }}
        ]
    }}

    Generate exactly {request.num_questions} practice questions and 10 quiz questions.
    Make questions relevant to {request.exam_type} exam pattern and {request.topic} topic.
    """

    completion = client.chat.completions.create(
        model="groq/compound",
        messages=[
            {
                "role": "system",
                "content": "You are an expert exam question generator. You must respond ONLY with valid JSON. No markdown, no code blocks, no explanations. Just pure JSON."
            },
            {
                "role": "user",
                "content": prompt
            }
        ],
        temperature=0.7,
        max_completion_tokens=2048,
        compound_custom={
            "tools": {
                "enabled_tools": ["web_search", "code_interpreter"]
            }
        }
    )

    raw_content = completion.choices[0].message.content.strip()
    
    # Clean up the response - remove markdown formatting if present
    if raw_content.startswith("```json"):
        raw_content = raw_content[7:]
    if raw_content.startswith("```"):
        raw_content = raw_content[3:]
    if raw_content.endswith("```"):
        raw_content = raw_content[:-3]
    
    raw_content = raw_content.strip()
    
    # Try to parse and validate the JSON
    try:
        parsed_data = json.loads(raw_content)
        
        # Validate structure
        if "practice_questions" not in parsed_data or "quiz_questions" not in parsed_data:
            raise ValueError("Missing required fields")
            
        # Ensure practice_questions has the right number
        if len(parsed_data["practice_questions"]) != request.num_questions:
            # Adjust if AI didn't follow the count
            parsed_data["practice_questions"] = parsed_data["practice_questions"][:request.num_questions]
            
        return {
            "exam_type": request.exam_type,
            "topic": request.topic,
            "questions": json.dumps(parsed_data)
        }
        
    except (json.JSONDecodeError, ValueError) as e:
        # Fallback: create a basic structure
        fallback_data = {
            "practice_questions": [
                {
                    "question": f"Sample question {i+1} for {request.topic}",
                    "options": ["Option A", "Option B", "Option C", "Option D"],
                    "correct_answer": "A",
                    "solution": f"Sample solution {i+1}"
                } for i in range(min(request.num_questions, 5))
            ],
            "quiz_questions": [
                {
                    "question": f"Quiz question {i+1} for {request.topic}",
                    "options": ["Option A", "Option B", "Option C", "Option D"],
                    "correct_answer": "A", 
                    "solution": f"Quiz solution {i+1}"
                } for i in range(10)
            ]
        }
        
        return {
            "exam_type": request.exam_type,
            "topic": request.topic,
            "questions": json.dumps(fallback_data)
        }
@app.post("/latest-trends")
async def latest_trends(request: TrendRequest):

    prompt = f"""
    User is studying {request.course}.

    IMPORTANT: Respond ONLY with valid JSON. Do not include any markdown formatting, code blocks, or extra text.

    Provide latest trends in this field including:
    1. Latest technologies in demand
    2. Trending domains
    3. Career opportunities
    4. Complete roadmap from beginner to advanced
    5. Tools and programming languages to learn
    6. Industry demand and future scope

    Respond ONLY with this JSON structure:
    {{
        "course": "{request.course}",
        "trending_domains": ["domain1", "domain2", "domain3"],
        "technologies": ["tech1", "tech2", "tech3"],
        "roadmap": [
            {{"step": "Step 1: Beginner", "description": "Foundations", "duration": "3-4 months"}}
        ],
        "tools": ["tool1", "tool2"],
        "career_roles": ["role1", "role2"],
        "future_scope": "details"
    }}
    """

    completion = client.chat.completions.create(
        model="groq/compound",
        messages=[
            {
                "role": "system",
                "content": "You are an expert career advisor. Respond ONLY with valid JSON. No markdown, no explanations."
            },
            {
                "role": "user",
                "content": prompt
            }
        ],
        temperature=0.7,
        max_completion_tokens=2048,
        compound_custom={
            "tools": {
                "enabled_tools": [
                    "web_search",
                    "visit_website",
                    "code_interpreter"
                ]
            }
        }
    )

    raw_content = completion.choices[0].message.content.strip()
    
    if raw_content.startswith("```json"):
        raw_content = raw_content[7:]
    if raw_content.startswith("```"):
        raw_content = raw_content[3:]
    if raw_content.endswith("```"):
        raw_content = raw_content[:-3]
    
    raw_content = raw_content.strip()
    
    try:
        parsed_data = json.loads(raw_content)
        if "course" not in parsed_data:
            parsed_data["course"] = request.course
        return {
            "course": request.course,
            "trends": json.dumps(parsed_data)
        }
    except (json.JSONDecodeError, ValueError):
        fallback_data = {
            "course": request.course,
            "trending_domains": ["AI & ML", "Cloud Computing", "Data Science", "DevOps", "Cybersecurity"],
            "technologies": ["Python", "Kubernetes", "Docker", "TensorFlow", "AWS"],
            "roadmap": [
                {"step": "Step 1: Beginner Foundations", "description": f"Learn basics of {request.course}", "duration": "3-4 months"},
                {"step": "Step 2: Intermediate Development", "description": f"Build projects in {request.course}", "duration": "3-4 months"},
                {"step": "Step 3: Advanced Specialization", "description": f"Master {request.course}", "duration": "3-4 months"},
                {"step": "Step 4: Expert Mastery", "description": f"Lead in {request.course}", "duration": "ongoing"}
            ],
            "tools": ["Git", "Docker", "VS Code", "Postman", "JIRA"],
            "career_roles": [f"{request.course} Developer", f"{request.course} Architect", "Technical Lead", "Manager"],
            "future_scope": f"{request.course} is in high demand with excellent career growth and competitive salaries globally."
        }
        return {
            "course": request.course,
            "trends": json.dumps(fallback_data)
        }
