ai-service/requirements.txt
fastapi
uvicorn
groq 
python-dotenv